create file AllCodeString
